+++ realm/realm-library/src/main/cpp/io_realm_internal_objectstore_OsAsyncOpenTask.cpp
@@ -26,6 +26,7 @@
 
 #include <shared_realm.hpp>
 #include <sync/async_open_task.hpp>
+#include <memory>
 
 using namespace realm;
 using namespace realm::jni_util;
@@ -44,23 +45,29 @@ JNIEXPORT jlong JNICALL Java_io_realm_internal_objectstore_OsAsyncOpenTask_start
         auto& config = *reinterpret_cast<Realm::Config*>(config_ptr);
 
         std::shared_ptr<realm::AsyncOpenTask> task = Realm::get_synchronized_realm(config);
-        task->start([task_obj=global_obj](realm::ThreadSafeReference<realm::Realm> realm_ref, std::exception_ptr error) {
+
+        auto deleter = [](jobject obj) {
+            jni_util::JniUtils::get_env(true)->DeleteGlobalRef(obj);
+        };
+        std::shared_ptr<_jobject> task_obj(env->NewGlobalRef(global_obj), deleter);
+        task->start([task=std::move(task_obj)](realm::ThreadSafeReference<realm::Realm> realm_ref, std::exception_ptr error) {
             JNIEnv* local_env = jni_util::JniUtils::get_env(true);
             if (error) {
                 try {
                     std::rethrow_exception(error);
-                } catch (const std::exception& e) {
+                }
+                catch (const std::exception& e) {
                     jstring j_error_msg = to_jstring(local_env, e.what());
-                    local_env->CallObjectMethod(task_obj, java_notify_error, j_error_msg);
+                    local_env->CallObjectMethod(task.get(), java_notify_error, j_error_msg);
                     local_env->DeleteLocalRef(j_error_msg);
                 }
             }
             else {
                 auto realm = Realm::get_shared_realm(std::move(realm_ref));
                 realm->close();
-                local_env->CallVoidMethod(task_obj, java_notify_realm_ready);
+                local_env->CallVoidMethod(task.get(), java_notify_realm_ready);
             }
-            local_env->DeleteGlobalRef(task_obj);
+            const_cast<std::shared_ptr<_jobject>&>(task).reset();
         });
         return reinterpret_cast<jlong>(&(*task));
     }